requirements.txt
flask
openai
python-telegram-bot==22.7